CMPSC enhancements in cooperation with Richard P.
1. Reduced the stack usage
Reducing the stack usage by placing variables for the stack like regs
and r? in the struct context. This reduces popping these values on the
stack on a function call.

2. Return with cc3 on interrupt pending
In the old days CMPSC returned after a fixed number of compression or
expansion. Now CMPSC continues until the end (in most cases) or an
interrupt pending. This is just like on the real thing and this will
probably provide a better responsive system.

git-svn-id: svn://svn.hercules-390.org/hercules/trunk@6456 969364d3-6d58-4e41-9446-b7e21338ea1b

Includes the following corrections:
git-svn-id: svn://svn.hercules-390.org/hercules/trunk@6457 969364d3-6d58-4e41-9446-b7e21338ea1b
git-svn-id: svn://svn.hercules-390.org/hercules/trunk@6474 969364d3-6d58-4e41-9446-b7e21338ea1b
git-svn-id: svn://svn.hercules-390.org/hercules/trunk@6475 969364d3-6d58-4e41-9446-b7e21338ea1b
git-svn-id: svn://svn.hercules-390.org/hercules/trunk@6476 969364d3-6d58-4e41-9446-b7e21338ea1b
git-svn-id: svn://svn.hercules-390.org/hercules/trunk@6477 969364d3-6d58-4e41-9446-b7e21338ea1b
git-svn-id: svn://svn.hercules-390.org/hercules/trunk@6564 969364d3-6d58-4e41-9446-b7e21338ea1b