Fix all-zero MAC bug in LCS cmd reply frame handling
git-svn-id: svn://svn.hercules-390.org/hercules/trunk@7341 969364d3-6d58-4e41-9446-b7e21338ea1b

This patch for Hercules 3.10 spinhawk corrects a bug causing unreliable
LCS connections resulting from the guest incorrectly being told
its MAC address is 00:00:00:00:00:00 due to a race condition.

Symptoms of this bug include a frequently non-functioning guest network
connection as soon as networking is enabled in the guest.  That is to say,
as soon as the guest enables TCPIP networking it is, nevertheless, still
unable to reach the host or anywhere else, and neither the host nor anyone
else is able to reach the guest, even though the guest's networking
configuration is correct.

Depending on the speed of the host this behavior can occur greater than 50%
of the time.  The only way to correct the problem is to try again and again
until it finally works.