MSA4 provides a register with a key where other keys can be encrypted.
This key comes with a verification pattern, to verify if the pseudo
(encrypted) key is still valid. This pattern showed the time when the
key was generated. I used this time to randomize. It is a security issue
to show this parameter.

We still randomize on time, but the last 8 bytes of the verification
pattern are also random numbers.

Bernard

git-svn-id: svn://svn.hercules-390.org/hercules/trunk@7451 969364d3-6d58-4e41-9446-b7e21338ea1b