fix handling of ltdl/dlopen
Nothing was setting hc_cv_have_dlopen, so initialize that based on the
dlopen tests that libtool does for us.

With that in place, we can set hc_cv_have_lt_dlopen based on the included
ltdl flag.  This lets people who have a *nix system with libdl avoid the
ltdl code entirely.

Signed-off-by: Mike Frysinger <vapier@gentoo.org>

the git repo does not build for me -- it fails when trying to build its own
copy of ltdl.c. my guess is because hercules is trying to build ltdl.c itself
instead of using the libltdl subdir.

but rather than investigate that, i wanted to fix a different problem (that was
reported in the past but never addressed): hercules shouldn't always be using
its own copy of libltdl. it should be able to use the system one or, even
better, use the standard -ldl that *nix systems already have.

thus i fixed the code to stop hardcoding to "yes use local ltdl" to respect the
configure flag --with-included-ltdl. and then i fixed the variable
hc_cv_have_dlopen which the configure script tests, but never actually sets
anywhere.
-mike