Prevent further warnings
The algorithm is agnostic to the type of format (1 or 2) it is using. Therefore the compiler thinks there is a possible error. I changed the code where sd2 will be initialized to NULL. Just for killing the compiler warnings and unfortunately one pico second slower ;-)

Also changed the copright statements to the correct year.