Eliminate warning: using the result of an assignment as a condition without parentheses
The clang compiler flags the use of the bitwise-OR-assignment operator within an if-statement. The expression must be enclosed in an additional set of parentheses to suppress the warning.

general2.c:214:34: warning: using the result of an assignment as a condition without parentheses [-Wparentheses]
            if (*dest2++ |= *source2++) cc = 1;
                ~~~~~~~~~^~~~~~~~~~~~~
general2.c:214:34: note: place parentheses around the assignment to silence this warning
            if (*dest2++ |= *source2++) cc = 1;
                         ^
                (                     )
general2.c:214:34: note: use '!=' to turn this compound assignment into an inequality comparison
            if (*dest2++ |= *source2++) cc = 1;
                         ^~
                         !=