Add data exception on above 127 ECE fetching
Page 7-156 of the POP DZ9ZR009 states:

During expansion of each character symbol, either
the characters in the symbol or the dictionary entries
representing characters of the symbol are counted,
and a data exception is recognized if this count
becomes too large. If the characters in the symbol
are counted, the count can reach at least 260 without
the exception being recognized. If the dictionary
entries representing characters of the symbol are
counted, the count can reach at least 127 without the
exception being recognized.

Character counting was already implemented. With this patch both situations, including CCE fetching counting, are tested and a data exception will occur whichever situation occurs first.

Please contact me at bernard.van.der.helm(at)runbox.com for any question.