Modify dasdload to set DIRF bit instead of DOS bit
Because dasdload builds a format 5 DSCB with no free space
it sets the DIRF bit in ds4vtoci to force the operating system
VTOC conversion routine to calculate the free space and update
the format 5 DSCB the first time the volume is accessed.
This was previously achieved by setting the DOS bit.