authorize_asn may use wrong ATO address after SIE translation
In an SIE environment with more than 4Gb of main storage
the guest 31-bits ATO address may translate to a 64-bits
absolute address in host main storage.

In this case, at line dat.h:235, this translation code

   SIE_TRANSLATE(&ato, ACCTYPE_SIE, regs);

may lead to a 64-bits ato value.

The problem is the declaration of the "ato" variable, at
line dat.h:206, which should be RADR instead of U32.