Incorrect instruction trace operand data and "u" command disassembles incorrect virtual page
When tracing instructions or single stepping, incorrect instruction
operand data can sometimes be displayed on the Hercules console.
The data is incorrect because it is from the wrong virtual page.
This is due to a stale page table entry being retrieved from a
temporary translation lookaside buffer in translate_addr() due in
turn to regs->tlbID being set to zero which appears to not be
valid. Before calling translate_addr(), a temporary copy is made of
part of the regs structure by copy_regs() but tlbID is not copied
(because it appears in the regs structure after regs_copy_end) and
the temporary tlbID is not initialised.

The "u" command is similarly affected in that it can end up
disassembling code from the wrong virtual page.

The "v" command is not affected as it bypasses the temporary TLB.

Data used for actual instruction execution is not affected because
regs->tlbID is never zero when this is being fetched.