Defects in PCC instruction
CMACs generated using FCs 1-3, 9-11, 18-20 and 26-28 are always wrong.
While FCs 1-3 and 9-11 consistently generate the same wrong CMACs from the
same input, FCs 18-20 and 26-28 generate arbitrarily wrong CMACs from the
same input.

The root cause is a logic error in the implementation of the last block
subkey generation algorithm, and, in addition, for FCs 18-20 and 26-28, a
parameter block length mismatch.

Fixed the faulty algorithm and the parameter block length mismatch, and
verified that the CMACs found in the examples in Appendix D of the
NIST-CSD Special Publication 800-38B are computed correctly. Additionally
created a lot of arbitrary CMACs using OpenSSL and verified that the
results are matching those computed by PCC.

[dyncrypt.patch 2016-04-05 21:30]