Defects in PCC and KM instructions
The XTS related functions (codes 50, 52, 58 and 60 of the KM and the PCC
instructions) deliver invalid results. Results are stable, except for PCC
function codes 58 and 60, which deliver arbitrary results. None of the
functions ever delivers a correct result.

The arbitrary results from PCC function codes 58 and 60 are caused by
memory overlay due to an invalid length (too short) of the parameter
block. Fixing this overlay makes the results stable, but equally invalid
as with the other function codes.

The invalid results come from a misinterpretation of the algorithm to use
for the GF(2128) multiplication. All diagrams and descriptions in the POP
suggest it is the same algorithm as used with function code 65 of the KIMD
instruction. However, it is not: For the GCM situation in KIMD the
specification found in NIST Special Publication 800-38D, dated November
2007, is relevant, while for the XTS situation in KM and PCC the
specification found in IEEE standard 1619-2007 is relevant. In fact, both
algorithms are the same (and thus the POP is correct). However, the bit
order of the arguments isn't the same in both standards, which isn't
exactly mentioned in the POP (one can find it in the prefix pages, once
one knows what to search for).

Fix the memory overlay in PCC and use the correct algorithm for the
GF(2128) multiplication in KM and PCC.

[dyncrypt_xts.patch 2016-04-21 21:15]