Activate OPTION_STRICT_ALIGNMENT to avoid segfault in -O3 build
segfault is caused because OPTION_STRICT_ALIGNMENT is turned off
in some cases in machdep.h (line 576)

If you remove the conditional if-satement and only leave

        #define OPTION_STRICT_ALIGNMENT

then the -03 option will not segfault. The reason is that in -O3
the option -ftree-loop-vectorize is enabled (this is the option
that causes the segfault in -O3). This option is quite advanced
and optimizes certain "for" loops among others.

The segfault actually occurs at: vstore.h line 803

        for ( ; n >= 8; n -= 8, dest += 8, src += 8)
            store_dw_noswap(dest,fetch_dw_noswap(src));

this gets optimized by the -ftree-loop-vectorize and segfaults
because the strict alignment option is not observed.

Posted by: Truda <gert_caers@yahoo.com> 15/07/2016 17:38