Ensure any sockets opened get closed on exit
When devinit or detach is used to reset or
close a 2703 (commadpt) device, some sockets
and/or pipes may get left open, leaving
listening sockets in particular dangling and
inaccessible, requiring a quit from Hercules
to clean up.  This fix ensures that sockets
and pipes are closed at device shutdown time.