Fix bug in 2703 (commadpt) devices
While waiting for a bind() to succeed, attempting
a devinit or detach or anything that results in a
command getting sent to the device thread can
result in the command getting ignored or the
device getting hung due to a read from the wrong
end of the command pipe.  Also, the wfd file
descriptor set is specified twice in a call to
select().  The second one should be xfd.  Both
issues are addressed by this fix.