-Wmaybe-uninitialized warning in tcpnje.c (#100)
Some compilers worry that TTBlength may be used without being initialised in
tcpnje.c.  They do not realise that it will have been given a value on a
previous trip around the loop before it gets used.  This fix initialises
TTBlength to 0 in an effort to keep the compiler from worrying needlessly.

Co-authored-by: Peter Coghlan <git@beyondthepale.ie>