Fix numerous -Wmaybe-uninitialized warnings in control.c
Some compilers produce lots of warnings
relating to aste possibly being used
uninitialized in control.c.  This
appears to be because aste is updated
by translate_asn() (in dat.h) but not
if an exception path is taken.  Back
in control.c, this results in a call
to program_interrupt() but the
compiler doesn't realise that calls to
program_interrupt() will not return.